Refactor to components
