+++ packages/bygger/src/components/FormMetaDataEditor/FormMetadataEditor.tsx
@@ -1,24 +1,12 @@
-import { Copy } from "@navikt/ds-icons";
-import { Alert, Button, Checkbox, Fieldset, Select, Textarea, TextField, ToggleGroup } from "@navikt/ds-react";
-import { makeStyles, useAppConfig } from "@navikt/skjemadigitalisering-shared-components";
-import {
-  DeclarationType,
-  formDiffingTool,
-  InnsendingType,
-  MottaksadresseData,
-  NavFormType,
-  signatureUtils,
-  TEXTS,
-} from "@navikt/skjemadigitalisering-shared-domain";
-import { Link } from "react-router-dom";
-import { v4 as uuidv4 } from "uuid";
-import useMottaksadresser from "../../hooks/useMottaksadresser";
-import useTemaKoder from "../../hooks/useTemaKoder";
-import copy from "../../util/copy";
-import SignatureComponent from "../layout/SignatureComponent";
-import EnhetSettings from "./EnhetSettings";
+import { Alert, Fieldset, Textarea, TextField } from "@navikt/ds-react";
+import { formDiffingTool, NavFormType, TEXTS } from "@navikt/skjemadigitalisering-shared-domain";
+import AddressFields from "./fields/AddressFields";
+import BasicFields from "./fields/BasicFields";
+import DeclarationFields from "./fields/DeclarationFields";
+import EnhetFields from "./fields/EnhetFields";
+import SignatureFields from "./fields/SignatureFields";
+import SubmissionFields from "./fields/SubmissionFields";
 import LabelWithDiff from "./LabelWithDiff";
-import SubmissionTypeSelect from "./SubmissionTypeSelect";
 import { FormMetadataError, UpdateFormFunction } from "./utils";
 
 type UsageContext = "create" | "edit";
@@ -32,231 +20,18 @@ interface Props {
 
 type BasicFormProps = Props & { usageContext: UsageContext };
 
-const useStyles = makeStyles({
-  row: {
-    display: "flex",
-    justifyContent: "space-between",
-  },
-});
-
 const BasicFormMetadataEditor = ({ form, publishedForm, onChange, usageContext, errors }: BasicFormProps) => {
-  const { featureToggles, config } = useAppConfig();
-  const { mottaksadresser, ready: isMottaksAdresserReady, errorMessage: mottaksadresseError } = useMottaksadresser();
-  const { temaKoder, ready: isTemaKoderReady, errorMessage: temaKoderError } = useTemaKoder();
   const diff = formDiffingTool.generateNavFormSettingsDiff(publishedForm, form);
-  const styles = useStyles();
   const {
-    title,
-    path,
-    properties: {
-      isTestForm,
-      skjemanummer,
-      tema,
-      downloadPdfButtonText,
-      innsending: innsendingFraProps,
-      ettersending,
-      ettersendelsesfrist,
-      mottaksadresseId,
-      declarationText,
-      enhetMaVelgesVedPapirInnsending,
-      enhetstyper,
-      descriptionOfSignatures,
-      descriptionOfSignaturesPositionUnder,
-      signatures,
-    },
+    properties: { downloadPdfButtonText, descriptionOfSignatures },
   } = form;
 
-  const addNewSignature = () =>
-    onChange({
-      ...form,
-      properties: {
-        ...form.properties,
-        signatures: [
-          ...signatureUtils.mapBackwardCompatibleSignatures(signatures),
-          {
-            label: "",
-            description: "",
-            key: uuidv4(),
-          },
-        ],
-      },
-    });
-
-  const addExistingSignature = (newSignature, index) =>
-    onChange({
-      ...form,
-      properties: {
-        ...form.properties,
-        signatures: signatureUtils.mapBackwardCompatibleSignatures(signatures).map((signatureObject, i) => {
-          if (index === i) {
-            return newSignature;
-          } else {
-            return signatureObject;
-          }
-        }),
-      },
-    });
-
-  const removeSignature = (signatureKey) => {
-    const mappedSignatures = signatureUtils.mapBackwardCompatibleSignatures(signatures);
-    if (mappedSignatures.length > 0) {
-      const updatedSignatures = mappedSignatures.filter((s) => s.key !== signatureKey);
-
-      onChange({
-        ...form,
-        properties: {
-          ...form.properties,
-          signatures: updatedSignatures,
-        },
-      });
-    }
-  };
-
-  const innsending = innsendingFraProps || "PAPIR_OG_DIGITAL";
-
-  const toAddressString = (address: MottaksadresseData) => {
-    const linjer = [address.adresselinje1];
-    if (address.adresselinje2) {
-      linjer.push(address.adresselinje2);
-    }
-    if (address.adresselinje3) {
-      linjer.push(address.adresselinje3);
-    }
-    return `${linjer.join(", ")}, ${address.postnummer} ${address.poststed}`;
-  };
-
   const basicFields = () => (
-    <>
-      {testSkjemaFields()}
-      <TextField
-        className="mb"
-        label="Skjemanummer"
-        type="text"
-        id="skjemanummer"
-        placeholder="Skriv inn skjemanummer"
-        value={skjemanummer}
-        readOnly={usageContext === "edit"}
-        onChange={(event) =>
-          onChange({ ...form, properties: { ...form.properties, skjemanummer: event.target.value } })
-        }
-        error={errors?.skjemanummer}
-      />
-      <TextField
-        className="mb"
-        label={<LabelWithDiff label="Tittel" diff={!!diff.title} />}
-        type="text"
-        id="title"
-        placeholder="Skriv inn tittel"
-        value={title}
-        onChange={(event) => onChange({ ...form, title: event.target.value })}
-        error={errors?.title}
-      />
-      {temaKodeFields()}
-    </>
+    <BasicFields onChange={onChange} diff={diff} form={form} errors={errors} usageContext={usageContext} />
   );
 
-  const testSkjemaFields = () => (
-    <>
-      <div className={`mb ${styles.row}`}>
-        <Checkbox
-          id="teststatus"
-          checked={!!isTestForm}
-          onChange={(event) =>
-            onChange({ ...form, properties: { ...form.properties, isTestForm: event.target.checked } })
-          }
-        >
-          Dette er et testskjema
-        </Checkbox>
-        <span>
-          Skjemadelingslenke
-          <Button
-            onClick={() => copy(`${config!.fyllutBaseUrl}/test/login?formPath=${path}`)}
-            icon={<Copy />}
-            title="Kopier skjemadelingslenke"
-            variant="tertiary"
-            size="xsmall"
-          />
-        </span>
-      </div>
-    </>
-  );
-
-  const temaKodeFields = () => (
-    <>
-      <Select
-        className="mb-4"
-        label={<LabelWithDiff label="Tema" diff={!!diff.tema} />}
-        id="tema"
-        disabled={!isTemaKoderReady}
-        value={temaKoder?.find((temaKode) => temaKode.key === tema)?.key || ""}
-        onChange={(event) => onChange({ ...form, properties: { ...form.properties, tema: event.target.value } })}
-        error={errors?.tema}
-      >
-        <option value="">{"Velg tema"}</option>
-        {temaKoder?.map(({ key, value }) => (
-          <option key={key} value={key}>
-            {`${value} (${key})`}
-          </option>
-        ))}
-      </Select>
-      {temaKoderError && (
-        <Alert variant="error" size="small">
-          {temaKoderError}
-        </Alert>
-      )}
-    </>
-  );
-
-  const summaryPageDeclarationFields = () => (
-    <>
-      <LabelWithDiff label="Erklæring på oppsummeringsside" diff={!!diff.declarationType} />
-      <div className="mb">
-        <ToggleGroup
-          size="small"
-          defaultValue={form.properties.declarationType ?? DeclarationType.none}
-          onChange={(value) => {
-            const declarationType = value as DeclarationType;
-            let properties;
-            if (declarationType !== DeclarationType.custom && form.properties.declarationText) {
-              properties = { ...form.properties, declarationType, declarationText: undefined };
-            } else {
-              properties = { ...form.properties, declarationType };
-            }
-
-            onChange({
-              ...form,
-              properties,
-            });
-          }}
-          className="mb-4"
-        >
-          <ToggleGroup.Item value={DeclarationType.none}>Ingen</ToggleGroup.Item>
-          <ToggleGroup.Item value={DeclarationType.default}>Standard</ToggleGroup.Item>
-          <ToggleGroup.Item value={DeclarationType.custom}>Tilpasset</ToggleGroup.Item>
-        </ToggleGroup>
-
-        {form.properties.declarationType === DeclarationType.custom && (
-          <Textarea
-            label={<LabelWithDiff label="Tilpasset erklæringstekst" diff={!!diff.declarationText} />}
-            value={declarationText || ""}
-            onChange={(event) =>
-              onChange({
-                ...form,
-                properties: { ...form.properties, declarationText: event.target.value },
-              })
-            }
-            error={errors?.declarationText}
-          />
-        )}
-
-        {form.properties.declarationType === DeclarationType.default && (
-          <div>
-            <label className="navds-label">Standard erklæringstekst</label>
-            <div>{TEXTS.statiske.declaration.defaultText}</div>
-          </div>
-        )}
-      </div>
-    </>
+  const declarationFields = () => (
+    <DeclarationFields onChange={onChange} diff={diff} form={form} errors={errors}></DeclarationFields>
   );
 
   const downloadPdfButtonTextFields = () => (
@@ -277,147 +52,12 @@ const BasicFormMetadataEditor = ({ form, publishedForm, onChange, usageContext, 
   );
 
   const submissionFields = () => (
-    <>
-      <SubmissionTypeSelect
-        name="form-innsending"
-        label={<LabelWithDiff label="Innsending" diff={!!diff.innsending} />}
-        value={innsending}
-        onChange={(event) =>
-          onChange({
-            ...form,
-            properties: { ...form.properties, innsending: event.target.value as InnsendingType },
-          })
-        }
-      />
-      <SubmissionTypeSelect
-        name="form-ettersending"
-        label={<LabelWithDiff label="Ettersending" diff={!!diff.ettersending} />}
-        value={ettersending}
-        allowEmpty={true}
-        error={errors?.ettersending}
-        onChange={(event) =>
-          onChange({
-            ...form,
-            properties: { ...form.properties, ettersending: event.target.value as InnsendingType },
-          })
-        }
-      />
-      {!!ettersending && ettersending !== "INGEN" && (
-        <TextField
-          className="mb"
-          label={<LabelWithDiff label="Ettersendelsesfrist (dager)" diff={!!diff.ettersendelsesfrist} />}
-          type="number"
-          id="ettersendelsesfrist"
-          value={ettersendelsesfrist || ""}
-          onChange={(event) =>
-            onChange({
-              ...form,
-              properties: { ...form.properties, ettersendelsesfrist: event.target.value },
-            })
-          }
-          placeholder={"Standard (14 dager)"}
-        />
-      )}
-      {innsending === "INGEN" && (
-        <>
-          <TextField
-            className="mb"
-            label={<LabelWithDiff label="Overskrift til innsending" diff={!!diff.innsendingOverskrift} />}
-            value={form.properties.innsendingOverskrift || ""}
-            onChange={(event) =>
-              onChange({
-                ...form,
-                properties: { ...form.properties, innsendingOverskrift: event.target.value },
-              })
-            }
-          />
-          <Textarea
-            className="mb"
-            label={<LabelWithDiff label="Forklaring til innsending" diff={!!diff.innsendingForklaring} />}
-            value={form.properties.innsendingForklaring || ""}
-            onChange={(event) =>
-              onChange({
-                ...form,
-                properties: { ...form.properties, innsendingForklaring: event.target.value },
-              })
-            }
-          />
-        </>
-      )}
-    </>
+    <SubmissionFields onChange={onChange} diff={diff} form={form} errors={errors}></SubmissionFields>
   );
 
-  const addressFields = () => (
-    <>
-      {(innsending === "KUN_PAPIR" || innsending === "PAPIR_OG_DIGITAL") && (
-        <div>
-          <Select
-            className="mb-4"
-            label={<LabelWithDiff label="Mottaksadresse" diff={!!diff.mottaksadresseId} />}
-            name="form-mottaksadresse"
-            id="form-mottaksadresse"
-            value={mottaksadresseId}
-            disabled={!isMottaksAdresserReady}
-            onChange={(event) =>
-              onChange({
-                ...form,
-                properties: {
-                  ...form.properties,
-                  mottaksadresseId: event.target.value || undefined,
-                  enhetMaVelgesVedPapirInnsending: false,
-                },
-              })
-            }
-          >
-            <option value="">
-              {mottaksadresseId && !isMottaksAdresserReady ? `Mottaksadresse-id: ${mottaksadresseId}` : "Standard"}
-            </option>
-            {mottaksadresser.map((adresse) => (
-              <option value={adresse._id} key={adresse._id}>
-                {toAddressString(adresse.data)}
-              </option>
-            ))}
-          </Select>
-          {mottaksadresseError && (
-            <Alert variant="error" size="small">
-              {mottaksadresseError}
-            </Alert>
-          )}
-        </div>
-      )}
-      <div className="mb">
-        <Link to="/mottaksadresser">Rediger mottaksadresser</Link>
-      </div>
-    </>
-  );
+  const addressFields = () => <AddressFields onChange={onChange} diff={diff} form={form}></AddressFields>;
 
-  const enhetsListFields = () => (
-    <>
-      {(innsending === "KUN_PAPIR" || innsending === "PAPIR_OG_DIGITAL") &&
-        !mottaksadresseId &&
-        featureToggles?.enableEnhetsListe && (
-          <div className="mb">
-            <EnhetSettings
-              enhetMaVelges={!!enhetMaVelgesVedPapirInnsending}
-              selectedEnhetstyper={enhetstyper}
-              onChangeEnhetMaVelges={(selected) =>
-                onChange({
-                  ...form,
-                  properties: {
-                    ...form.properties,
-                    enhetMaVelgesVedPapirInnsending: selected,
-                    enhetstyper: selected ? form.properties.enhetstyper : undefined,
-                  },
-                })
-              }
-              onChangeEnhetstyper={(enhetstyper) =>
-                onChange({ ...form, properties: { ...form.properties, enhetstyper } })
-              }
-            />
-          </div>
-        )}
-    </>
-  );
+  const enhetsListFields = () => <EnhetFields onChange={onChange} form={form}></EnhetFields>;
 
   const instructionFields = () => (
     <>
@@ -435,36 +75,7 @@ const BasicFormMetadataEditor = ({ form, publishedForm, onChange, usageContext, 
     </>
   );
 
-  const signatureFields = () => (
-    <>
-      <Checkbox
-        className="mb"
-        checked={!!descriptionOfSignaturesPositionUnder}
-        onChange={(event) =>
-          onChange({
-            ...form,
-            properties: { ...form.properties, descriptionOfSignaturesPositionUnder: event.target.checked },
-          })
-        }
-      >
-        Plasser under signaturer
-      </Checkbox>
-      {signatureUtils.mapBackwardCompatibleSignatures(signatures)?.map((signature, index) => (
-        <div key={signature.key}>
-          <SignatureComponent
-            signature={signature}
-            diff={diff.signatures?.[signature.key]}
-            index={index}
-            onChange={(newSignature) => addExistingSignature(newSignature, index)}
-            onDelete={() => removeSignature(signature.key)}
-          />
-        </div>
-      ))}
-      <Button variant="secondary" className="mb" onClick={addNewSignature}>
-        Legg til signatur
-      </Button>
-    </>
-  );
+  const signatureFields = () => <SignatureFields onChange={onChange} diff={diff} form={form} />;
 
   return (
     <Fieldset hideLegend legend="">
@@ -472,7 +83,7 @@ const BasicFormMetadataEditor = ({ form, publishedForm, onChange, usageContext, 
       <div className="mb">{basicFields()}</div>
       {usageContext === "edit" && (
         <>
-          {summaryPageDeclarationFields()}
+          {declarationFields()}
           {downloadPdfButtonTextFields()}
           {submissionFields()}
           {addressFields()}
