+++ packages/shared-domain/src/index.ts
@@ -39,7 +39,7 @@ import MockedComponentObjectForTest from "./summary/MockedComponentObjectForTest
 import TEXTS from "./texts";
 import dateUtils from "./utils/date";
 import featureUtils, { FeatureTogglesMap } from "./utils/featureUtils";
-import formDiffingTool from "./utils/formDiffingTool";
+import formDiffingTool, { NavFormSettingsDiff } from "./utils/formDiffingTool";
 import navFormioUtils from "./utils/formio";
 import { guid } from "./utils/guid";
 import localizationUtils from "./utils/localization";
@@ -109,4 +109,5 @@ export type {
   UkjentBruker,
   ReportDefinition,
   Operator,
+  NavFormSettingsDiff,
 };
